+++ src/app/common/validation/forms/address-validators.ts
@@ -9,7 +9,7 @@ import { validateAddressChain, validateStacksAddress } from '@app/common/stacks-
 export function btcAddressValidator() {
   return yup
     .string()
-    .defined()
+    .defined('Enter a bitcoin address')
     .test((input, context) => {
       if (!input) return false;
       if (!validate(input))
@@ -20,11 +20,18 @@ export function btcAddressValidator() {
     });
 }
 
-export function btcAddressNetworkValidatorFactory(network: NetworkModes) {
+function btcAddressNetworkValidatorFactory(network: NetworkModes) {
   return (value?: string) => {
     if (!isString(value)) return false;
     return validate(value, network as Network);
   };
+}
+
+export function btcAddressNetworkValidator(network: NetworkModes) {
+  return yup.string().test({
+    test: btcAddressNetworkValidatorFactory(network),
+    message: 'Bitcoin address targets different network',
+  });
 }
 
 export function stxAddressNetworkValidatorFactory(currentNetwork: NetworkConfiguration) {
