+++ src/app/common/validation/forms/fee-validators.ts
@@ -6,12 +6,10 @@ import { Money } from '@shared/models/money.model';
 import { isNumber } from '@shared/utils';
 
 import { formatInsufficientBalanceError, formatPrecisionError } from '@app/common/error-formatters';
-import {
-  btcCurrencyAmountValidator,
-  stxCurrencyAmountValidator,
-} from '@app/common/validation/forms/currency-validators';
+import { btcAmountPrecisionValidator } from '@app/common/validation/forms/currency-validators';
 
 import { btcToSat, moneyToBaseUnit, stxToMicroStx } from '../../money/unit-conversion';
+import { stxAmountValidator } from './amount-validators';
 
 interface FeeValidatorFactoryArgs {
   availableBalance?: Money;
@@ -39,7 +37,7 @@ export function btcFeeValidator(availableBalance?: Money) {
   return feeValidatorFactory({
     availableBalance,
     unitConverter: btcToSat,
-    validator: btcCurrencyAmountValidator,
+    validator: btcAmountPrecisionValidator,
   });
 }
 
@@ -47,6 +45,6 @@ export function stxFeeValidator(availableBalance?: Money) {
   return feeValidatorFactory({
     availableBalance,
     unitConverter: stxToMicroStx,
-    validator: stxCurrencyAmountValidator,
+    validator: stxAmountValidator,
   });
 }
