+++ src/app/pages/send/send-crypto-asset-form/form/stx/stx-crypto-currency-send-form.tsx
@@ -21,7 +21,10 @@ import { useDrawers } from '@app/common/hooks/use-drawers';
 import { useRouteHeader } from '@app/common/hooks/use-route-header';
 import { convertAmountToBaseUnit } from '@app/common/money/calculate-money';
 import { useWalletType } from '@app/common/use-wallet-type';
-import { stxAmountValidator } from '@app/common/validation/forms/amount-validators';
+import {
+  stxAmountValidator,
+  stxAvailableBalanceValidator,
+} from '@app/common/validation/forms/amount-validators';
 import { stxFeeValidator } from '@app/common/validation/forms/fee-validators';
 import { stxMemoValidator } from '@app/common/validation/forms/memo-validators';
 import {
@@ -57,7 +60,7 @@ import { SelectedAssetField } from '../../components/selected-asset-field';
 import { SendAllButton } from '../../components/send-all-button';
 import { StacksRecipientField } from '../../family/stacks/components/stacks-recipient-field';
 import { useSendFormNavigate } from '../../hooks/use-send-form-navigate';
-import { createDefaultInitialFormValues } from '../../send-form.utils';
+import { createDefaultInitialFormValues, defaultFormikProps } from '../../send-form.utils';
 
 export function StxCryptoCurrencySendForm() {
   const navigate = useNavigate();
@@ -100,7 +103,7 @@ export function StxCryptoCurrencySendForm() {
   });
 
   const validationSchema = yup.object({
-    amount: stxAmountValidator(availableStxBalance),
+    amount: stxAmountValidator().concat(stxAvailableBalanceValidator(availableStxBalance)),
     recipient: stxRecipientValidator(currentAccountStxAddress, currentNetwork),
     recipientAddressOrBnsName: stxRecipientAddressOrBnsNameValidator({
       client,
@@ -135,10 +138,8 @@ export function StxCryptoCurrencySendForm() {
     <Formik
       initialValues={initialValues}
       onSubmit={async (values, formikHelpers) => await previewTransaction(values, formikHelpers)}
-      validateOnBlur={false}
-      validateOnChange={false}
-      validateOnMount={false}
       validationSchema={validationSchema}
+      {...defaultFormikProps}
     >
       {props => (
         <NonceSetter>
