+++ src/app/pages/send/send-crypto-asset-form/send-form.utils.ts
@@ -1,3 +1,8 @@
 export function createDefaultInitialFormValues<T extends Record<string, unknown>>(obj: T) {
   return { amount: '', recipient: '', ...obj };
 }
+
+export const defaultFormikProps = {
+  validateOnChange: false,
+  validateOnMount: false,
+} as const;
